refactor routes to seperate files for cleaner code
